verity: do all hashing in dm-bht.c
We were doing the hashing of the disk blocks in dm-verity.c and hashing
hash blocks in dm-bht.c. We can simplifiy the code by doing all the
hashing in dm-bht.c. We couldn't do this earlier because we had
to handle unaligned reads (no longer an issue).

I also removed the BUG_ON in dm_bht_get_node because I wanted to use it
in dm_bht_check_block with depth == bht->depth.

The kernel piece has already been reviewed and submitted. What is different
in this CL are the modifications to the tests.

BUG=9752
TEST=Ran dm-verity.git unit tests. Ran platform_DMVerityCorruption on H/W.
kernel.git Review URL: http://codereview.chromium.org/6626037

Change-Id: Ic817a53f383f7727742c77c2064cdb0afe595c43

R=wad@chromium.org,taysom@chromium.org,ups@chromium.org

Review URL: http://codereview.chromium.org/6695038