verity: remove short-circuiting of verify_path
This removes the optimization which short-circuited verify_path. The net
result is the addition of a few cycles and the removal of a few so zero
or negligible cycles added.

The main benefit is the removal of the one of the larget comments in the
code. Ran BootPerfServer to verify no noticable overhead added.

Before:

  seconds_power_on_to_login                                       8.44

After:

  seconds_power_on_to_login                                       8.4

This result is within variance.

BUG=9752
TEST=Ran dm-verity.git unit tests. Ran platform_DMVerityCorruption on H/W.

Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
kernel.git Review URL: http://codereview.chromium.org/6659004

TBRing because this has already been LGTMed and committed to kernel.git.

Change-Id: Ib9384f3a5addc014f679cc1d46638bf782b12c9f

TBR=wad@chromium.org,taysom@chromium.org,ups@chromium.org

Review URL: http://codereview.chromium.org/6725015