verity: short-circuit dm_bht_populate
If a node is verified, so are all its parents. This is because we
verify top down. So we can break as soon as we see a verfied node.
Its parents are already verified and populated.

BUG=chromium-os:9752
TEST=Ran dm-verity.git unit tests. Ran platform_DMVerityCorruption on H/W.

Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
kernel.git Reviewed-on: http://gerrit.chromium.org/gerrit/4452

Code has already been reviewed and committed to kernel.git.

Change-Id: I9b6a1f104b62d4159c96ccdb466181f2c018c817
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6774