verity: use key-value arguments.
This opens the way for salting and other such improvements to dm-verity.

BUG=chromium-os:15772
TEST=Adhoc
Make sure you have rebuilt verity both for host and target!
Build a regular image for Arm (I used Kaen). Ensure it boots.
Build a recovery image for Arm. Ensure it boots and completes recovery; when you
remove the stick, ensure the resulting system is useable.
Build a regular image for x86 (I used Alex). Ensure it boots.
Build a recovery image for x86. Ensure it boots and completes recovery.
Build a vm image for x86. Ensure it boots.

Change-Id: I184d55ff7ccaf1612a1eda7ea1605a4302be6595
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/6085
Reviewed-by: Will Drewry <wad@chromium.org>