verity-static: bring back static compilation for verity
This common.mk is old, but the basics of this are good. This shows how to get
all your static libraries dynamically and link them.  The trick is picking PIC
over PIE since toolchains rarely ship with PIE components.  If we havepie
archives, we should use them :)

TEST=builds and runs
BUG=none

Change-Id: I070d4eff8fb3bdc604ea67ee3f67dac50d03740e
Reviewed-on: https://gerrit.chromium.org/gerrit/14631
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>
Commit-Ready: Will Drewry <wad@chromium.org>