Create a static library of libdm-bht
This pulls in the C sources into a library as well as the kernel
implementations of the hashes, so that the installer can statically
link them in.  We have to do a few non-obvious things in order to get
libdm-bht.a instead of libdm-bht.pic.a due to the double invocation of
that rule -- added TODO(wad) to fix that.

BUG=chromium-os:25377
TEST=manual: build verity, ensure libdm-bht.a is created (won't be
installed)

Change-Id: I4037c68d3fc2c1dfdcfed4a603b49b57dd06562b
Reviewed-on: https://gerrit.chromium.org/gerrit/15838
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>