shill: PowerManager: Track un-registered SuspendDelays
In R37 and R38 we added a change to shill to add a SuspendDelay
for the manager as it starts up (CL:206596).  Unfortunately this
introduced a race where if powerd was not running at boot before
shill had started, it was possible for this registration to fail.
Further exacerbating this issue, if PowerManager fails the call
to RegisterSuspendDelay on the PowerManagerProxy (e.g., if the
proxy does not exist) it does not take note of the SuspendDelay.

As a result, if powerd isn't running at shill startup, the
Manager will never be notified about suspend and resume events.
R39 and newer don't have this problem since the code has been
simplified quite a bit to remove the SuspendDelay map, and is
configured with a single suspend and resume callback.

This CL fixes the problem in R38 in as minimal a way as possible.
SuspendDelay entries are added regardless of the success or
failure of the RegisterSuspendDelay call, but the entry is now
annotated to indicate whether the SuspendDelay was successfully
registered.  Events are delivered to all SuspendDelays whether
or not they are registered, but calls to ReportSuspendReadiness
are ignored.

This CL maintains the behavior of the original CL, where suspend
delays that fail to re-register continue to perform as if they
were (notably ReportSuspendReadiness will be called if we ever
received a value from a previous registration).

BUG=chromium:411370
TEST=Manual: Observe error message from PowerManagerProxy on
startup in calling RegisterSuspendDelay, but on resume from
suspend note that wifi.cc reports OnAfterResume has been called.

Change-Id: Ia56c2a6e6b8a0ccaf3f49ba6ba65011276dce40b
Reviewed-on: https://chromium-review.googlesource.com/216852
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>