init: Preserve .labmachine across powerwash for debug images.
In test images, certain important behaviors are enabled if a file
named .labmachine is present in the root of the stateful file
system.  The behaviors are needed to recover from various critical
events, such as loss of network, and loss of AC power.

The file doesn't automatically exist; it must be created by the
Autotest install code.  Additionally, power wash can delete the
file.  This changes clobber-state to preserve the file across
powerwash in test and dev images.  Without this change, devices may
become unrepairable after certain tests.

BUG=chromium:463177
TEST=manually powerwash with the .labmachine file present

Change-Id: I3f74bf923e74bc1993ea058db0fc50cc1970ebff
Reviewed-on: https://chromium-review.googlesource.com/260581
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>