shill: Replace Wake on WiFi "disabled" log messages
Replace error messages stating that Wake on WiFi is "disabled" with error
messages that say that Wake on WiFi is "not supported" in all cases where Wake
on WiFi-related functionality when the DISABLE_WAKE_ON_WIFI compile-time flag is
set. Do this so that error messages are consistent when the platform does not
supported Wake on WiFi (i.e. DISABLE_WAKE_ON_WIFI is set).

Also
- Log "Wake on WiFi not supported" messages in the WiFi scope at level 7, since
  we consider running into this error "normal" behavior by Chrome and do not
  want to spam the logs at the default level.
- Remove TODO from comment above |wake_on_wifi_features_enabled_|
  initialization since Chrome is responsible for setting this value via DBus.

BUG=chromium:463954
TEST=Compile shill and run unit tests with and without the "wake_on_wifi" USE
flag.

Reviewed-on: https://chromium-review.googlesource.com/255783
Tested-by: Samuel Tan <samueltan@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>
Change-Id: Ic09a71861d7eda37d67c2320d4262d0a13187188
Reviewed-on: https://chromium-review.googlesource.com/259813
Reviewed-by: Samuel Tan <samueltan@chromium.org>