cryptohome: Cache the EK public key before TPM ownership.
Before the TPM is owned it is possible to read the EK public key without
owner permissions. Caching this value allows applications to query for
this information without waiting for the TPM ownership flow to complete.

This is a very manual cherry-pick of CL:261631.

BUG=brillo:638
TEST=FEATURES=test emerge-amd64-generic cryptohome

Change-Id: I8f5b112e66d0eb7a311ae542909d03125d270791
Reviewed-on: https://chromium-review.googlesource.com/262940
Commit-Queue: Garret Kelly <gdk@chromium.org>
Trybot-Ready: Garret Kelly <gdk@chromium.org>
Tested-by: Garret Kelly <gdk@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>