init: send-kernel-errors: Add kernel error reporting via UMA
On occasion we may want to scan the kernel's startup log for specific
strings that indicate faults during the boot process.  We already
have a warn_collector daemon that constantly scans the kernel log
in the background, but it isn't well-suited for this task for three
reasons:

 - It monitors the log file (on disk), so it doesn't have an easy way
   to tell where the current boot started.  It gets around this
   problem by ignoring all messages that were logged before the daemon
   was started.  So as it stands, it won't catch any errors that pop
   up during boot.

 - It is designed to look for standard Linux WARN_ON-style messages,
   which include a crash signature and backtrace.

 - In many cases we need to be able to find the ratio of good:bad
   units, not just the raw number of occurrences.

So, the most straightforward solution is to add another send-* upstart
job that calls metrics_client for each string that we want to monitor.

BUG=chromium:458878
TEST=boot test with both good and bad results, then peek into UMA to
see stats

Change-Id: Ic8e30d7a52e6622e2d555ef818fe11e132046097
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/272278
(cherry picked from commit de2c6dc85ae7aaedd63db448f14db67389da0786)
Reviewed-on: https://chromium-review.googlesource.com/275694