buffet: add a timeout to HTTP requests
CloudCommandProxy makes the understandable assumption that either one of
the success or error callbacks it provides to
chromeos::http::SendRequest will be invoked in the future.  When neither
of these callbacks are invoked future updates to that command will not
be sent to the cloud.  Add a timeout to all requests made by buffet to
allow these assumptions to function.

BUG=chrome-os-partner:40663
TEST=FEATURES=test emerge-amd64-generic test

Change-Id: Ia1e1f265818c9c69427fe641651c76c6783d720f
Reviewed-on: https://chromium-review.googlesource.com/274046
Reviewed-by: Garret Kelly <gdk@chromium.org>
Commit-Queue: Garret Kelly <gdk@chromium.org>
Tested-by: Garret Kelly <gdk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/274515
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>