buffet: Handle cloud request failures due to device deleted on cloud
When making cloud request and receiving "authorization failed" error,
we are refreshing the access token. If that fails with "invalid_grant",
we are certain that the device's robot account has been deleted, so
we should treat the device as no longer registered with the cloud, and
delete its cached credentials.

This CL adds OnDeviceDeleted() notification which wipes out the device
registration information. This notification is called when refreshing
the access token fails with "invalid_grant".

A follow-up CL will add handling for DEVICE_DELETED XMPP notification.

Merging CL:281393 to M44 branch

BUG=brillo:1215
TEST=`FEATURES=test emerge-link buffet`
     `test_that -b link 100.96.49.59 e:buffet_.*`
CQ-DEPEND=CL:281815

Change-Id: I615cd0f051a336ce80f10d8830af2f7dce5d32df
Reviewed-on: https://chromium-review.googlesource.com/281961
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>