power: fix up reading of battery attributes
This reads the energy_now attribute for the battery_energy value if it exists.
Before, we would multiply the charge by the nominal voltage even if the
attribute was there. It also uses the nominal voltage instead of the current
voltage when calculating charge from energy to be consistent. This modifies a
unittest for check this behavior and adds another to check that energy_now is
read for the battery_energy value.

BUG=chrome-os-partner:37123
TEST=run power_supply_info

Change-Id: I436b0a1394f2de8d6255387578b7471d2d407209
Reviewed-on: https://chromium-review.googlesource.com/293622
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
(cherry picked from commit fd7768a5fd8419474bcd047149ecd4f3c0bf5008)
Reviewed-on: https://chromium-review.googlesource.com/300951
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>