shill: stop IcmpSession before calling result callback
Stop IcmpSession before calling the result callback in
IcmpSession::ReportResultAndStopSession, since the
result callback might delete the IcmpSession object itself,
which will causes a segfault when we return to
IcmpSession::ReportResultAndStopSession and attempt to
call the Stop() member function of the deleted object.

To accomodate to this change, do not reset the result
callback in IcmpSession::Stop. This is safe as stop resets
all callbacks that might invoke the result callback again.

BUG=chromium:526165
TEST=Compile shill and run unit tests.
TEST=Modified shill so I can manually trigger ConnectionDiagnostics on a DNS
     timeout, and verified that diagnostics complete successfully, and shill
     does not crash.

Change-Id: Ic3804a2d852b13cd31a898ddbf7a857edac58b74
Reviewed-on: https://chromium-review.googlesource.com/296037
Reviewed-by: Samuel Tan <samueltan@chromium.org>
Tested-by: Samuel Tan <samueltan@chromium.org>