shill: don't truncate sequence numbers
Both when handling NetlinkPackets and when managing the pending dump
sequence number shill has previously truncated the 32bit netlink
sequence to a uint16_t.  This causes issues where the untruncated
sequence is used as the key in the map of handlers, meaning that the
handler never gets invoked and the messages are then treated as if they
were broadcast messages.

BUG=chrome-os-partner:46556
TEST=created a program that issued more than 2^16-1 dump requests

Change-Id: I8360802807e5868b9bd692266713d74896d2a11b
Reviewed-on: https://chromium-review.googlesource.com/307110
Reviewed-by: Garret Kelly <gdk@chromium.org>
Commit-Queue: Garret Kelly <gdk@chromium.org>
Tested-by: Garret Kelly <gdk@chromium.org>