libcontainer: copy_minor: Do not fail when host device does not exist
It is useful to be able to specify -1 as minor number in runtime.json,
so that the minor number is copied from outside the container.

However, the device may not exist outside of the container, in that case,
just skip it and not create it inside the container.

BUG=chrome-os-partner:54825
BUG=b:31090263
TEST=HWTest on cyan, veyron_minnie

Change-Id: I7d4c17a91cd421765a1764bc3b9fcfa64152e0d1
Previous-Reviewed-on: https://chromium-review.googlesource.com/357701
(cherry picked from commit 69be6a0d06350ab40ffe459c992b1f962642be7c)
Reviewed-on: https://chromium-review.googlesource.com/376459
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>