login: Exception added for parrot and glimmer when VPD update fails
These types of devices will be able to be enrolled even when VPD
update fails. This change doesn't not make insecure the
parrot/glimmer devices for which VPD update works well. Only the
devices where the operation fails become insecure and the only
indication of that can be seen in the logs.

BUG=chromium:653814
TEST=manually tested on parrot, glimmer and reks devices.

Change-Id: I834a63000d827f9aa38b2a589e5f65d88f5d5db0
Reviewed-on: https://chromium-review.googlesource.com/400238
Reviewed-by: Dan Erat <derat@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>