vm_tools: seneschal: Force unmount MNT_DETACH for MyFiles if EBUSY
MyFiles has Downloads bind-mounted within it.  When MyFiles is shared
and mounted, the Downloads mount propagates.

When we unshare MyFiles, we first try to unmount MyFiles/Downloads, but
it fails with EINVAL.  If we then continue and unmount MyFiles, it fails
with EBUSY regardless of whether any files are open.

This fixes by ignoring the first error for MyFiles/Downloads, and then
reattempting to unmount with MNT_DETACH set if unmounting MyFiles fails
with EBUSY.

BUG=chromium:1112190
TEST=share then unshare MyFiles root

Change-Id: Ib3995d8f867928c451eab3ff987b391cce795f84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1404454
Tested-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>