shill: key_value_store: Add "Variant" to {Contains,Get,Set} methods
Following CLs will templatize the KeyValueStore Contains*, Get*, and
Set* methods. As the identifiers 'Contains', 'Get', and 'Set' will
then be taken, we add a "Variant" suffix to the existing Contains,
Get, and Set methods. Future CLs could make template specializations
for the brillo::Any type and avoid having separate *Variant methods if
that is so desired.

This is a reland of CL:1652470.

BUG=chromium:936111
TEST=All unit tests are passing.

Change-Id: I6b67f409c97d7639d30714ccd8e0c62c90d3f20f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2090064
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>