vm_tools: add disk resize to concierge and maitred
Add a mechanism to resize the disk image and the filesystem on it via
concierge RPCs.

This is accomplished by extending maitred with a new pair of RPCs to
resize the btrfs filesystem on the stateful disk, plus a new pair of
RPCs in concierge to manage the overall resize process.  Concierge
manages the sequencing of operations, issuing a filesystem resize and
disk image resize (via `crosvm disk resize`) in the correct order
depending on whether the resize operation is shrinking or expanding the
disk.

BUG=chromium:858815
TEST=`vmc resize termina <size>`, expand and shrink disk

Change-Id: I6bbb5f1a4f77c860013f46bf1ed0716579fc0c84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2015622
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>