trunks: fix parsing response handles
Before this CL, if a response contained payload of size smaller than
required to hold all handles expected by the command, the missing
handles were set to 0 but the parsing continued. That led to errors
when attempting to replace those handles in the response with virtual
handles in resource manager.

This CL returns an empty vector of handles if an error occured while
parsing them. The parsing logic after extracting the handles already
checks that the expected number was returned in that vector, so parsing
short-circuits and safely returns RC_SIZE at this point, rather than
continuing to process the response.

BUG=chromium:910926
TEST=unit tests; test from the BUG

Cq-Depend: chromium:2152255
Change-Id: Icc07166db65e24cf9bf542e8bc870fba1f7ba99e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1776570
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Andrey Pronin <apronin@chromium.org>