cryptohome: Fuzzer for TPM 1.2 CMK parser
Add fuzzer for the code that performs parsing of the migrated
Certified Migratable Key on TPM 1.2.

This code is one of the most convoluted and rich for low-level
data parsing parts of the SignatureSealingBackend
implementation, which is used for challenge-response mounts
(a.k.a. smart card login).

BUG=chromium:837253
TEST=run locally

Change-Id: If86002384f9678514fbbb21620fca2f44041c80d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1872062
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>