mount-encrypted: Dumpe2fs on failing to mount the dm-crypt device
On failure to mount the dm-crypt device, use dumpe2fs to get
information about the superblock. Note that this does not
collect information about block groups.

The reasoning behind collecting this data from within
mount-encrypted is that the device is immediately teared down
on failure to mount.

BUG=chromium:1025044
TEST=use clobber collector to collect dumpe2fs logs.

Change-Id: Ib3eef8bcbb40da57aa3c5d17fe8875eea6f0eca0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1977850
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>