tpm_manager: constructor delegate of TpmManagerService
Before this CL, the constructor for production and testing are
respectively implemented. This makes the common construction for both
cases difficult to maintain.

This CL tackles this issue and now the common setup doesn't have to be
duplicated.

Note that the inline member initialization is not removed so we can make
sure we never have garbage values in those pointers.

BUG=b:147759766
TEST=build OK.

Change-Id: Idda794a9dd9cc7da0d5c106c66966509aeb05232
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2035197
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: John L Chen <zuan@chromium.org>