tpm_manager: rewrite state machine of TpmStatusImpl
Before this CL, the owner password status is invalidated so next time
when we get tpm status we still have to go through the entire flow.

This CL takes advantage of the information when the owner password is
set to a random one. This way, we can save several TPM calls, like
checking ownership, and default owner auth test.

Also, |TestTpmWithDefaultOwnerPassword| seems to be used only for
TPM 1.2; thus, removes it from the public interface of |TpmStatus|.

BUG=b:147759766
TEST=unittests for TPM1.2 and TPM2.0
TEST=tast run $DUT_IP hwsec.RetakeOwnership

Change-Id: Id29cb9065d1756be84b65c1f8a79d1a19d8fb2b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2036857
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>