cryptohome: check PCR0 state for quoting and DB sealing (monolithic)
When we encrypt attestation DB, the encryption key is sealed to PCR0.
If the PCR0 has been double-extended, we may not be able to recover
the key afterwards. Similarly, quoting PCR0 at a bad PCR0 state will
cause attestation to fail.

This CL checks if PCR0 is in the correct state when
  1. starting preparing for enrollment,
  2. generating the attestation DB sealing key, and
  3. quoting PCR0 when creatig AIK or migrating the attestation DB.

A similar change for distributed mode will come later.

BUG=b:138324811
TEST=unit tests;
     manually extend PCR0 on a chell before sealing the attestation DB
     and check if PrepareForEnrollment fails as expected; similar tests
     for quoting PCR0.

Change-Id: Ibe3e2d0f63ef3dc09da8f4e41b83c9e0ce1a323a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2011487
Commit-Queue: Wei-Cheng Xiao <garryxiao@chromium.org>
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>