ippusb_manager: Introducing keep-alive protocol
This change introduces the use of a 'keep-alive' protocol which is used
eliminate the possibility of the race condition caused when a print
request comes in when ippusbxd is in the process of shutting down.

Now, instead of just checking for the presence of <vid>_<pid>.sock,
ippusb_manager will issue an 'keep-alive' message to determine if
ippusbxd can handle print requests. If ippusbxd successfully
acknowledges the 'keep-alive' then ippusb_manager will opt not to spawn
a new instance of ipusbxd. Otherwise, it will wait for all of the
ippusbxd sockets to close before launching a new process.

BUG=chromium:1047022
TEST=Tested manually using virtual-usb-printer

Cq-Depend: chromium:2032665
Change-Id: Iae7b61a9094b1e1bf5cc3da54645f0a39067543b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2032705
Tested-by: David Valleau <valleau@chromium.org>
Commit-Queue: David Valleau <valleau@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>