shill: don't fail to probe cellular device with no IFLA_ADDRESS
It's possible for net interfaces associated with cellular modems
not to have a MAC address because they don't use the same MAC
address scheme. For many modems, the MAC address is randomly
generated and not stable, so we get past this check even if the
MAC is meaningless, but for some modem setups it is not present
at all.

BUG=b:146390732
TEST=verify device is registered when iface comes up on trogdor

Change-Id: Ife22987d34008ad48a2ba6aa21d7f7999f77ce08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2054768
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>