cros-disks: RarManager can access Play Files
Running rar2fs as part of the android-everybody group allows it to read
and mount RAR archives stored in Play Files.

Note that, since RarManager is not used by cros-disks yet, there is no
change in cros-disks's behavior. This is essentially preparing
RarManager to be a replacement of ArchiveManager.

BUG=chromium:996549
TEST=cros_workon_make cros-disks --test

Cq-Depend: chromium:2038216
Change-Id: Iab56ef970ef0791f1286e6a480618d537be24d5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2038220
Tested-by: François Degros <fdegros@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>