arc: networkd: Add bcast addr to ip addr add call.
This patch explicitly adds the broadcast address to the bridge
interface on creation. This used to happen automatically with ifconfig
but that was removed in https://crrev.com/c/2024209. The bcast addr is
calculable from the base addr and prefix length so probably nothing is
misbehaving but the CtsBionicTestCases.ifaddrs#getifaddrs_INET testcase
doesn't like this being 0.0.0.0.

BUG=b:149171301
TEST=units, flashed atlas verified bcast addr on arc bridges
TEST=test_that atlas cheets_CTS_P.9.0_r10.x86.CtsBionicTestCases PASSED

Change-Id: I3a16ecb43833d8accccda939b955f299814a75e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2045919
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>