chromeos-config: Use brillo::MkdirRecursively instead of base::CreateDirectory*
base::CreateDirectory and base::CreateDirectoryAndGetError are
hard-coded to create directories with permissions 0700, modulo the
umask. This was done in the early days of Chrome, before we had to
consider other users and the context of an operating system.

libbrillo provides MkdirRecursively, allowing us to specify the mode
of new directories. Use this function instead so we can make them with
0755 instead of 0700.

BUG=chromium:1041544
TEST=cros_config --mount_fallback $(mktemp -d), observe ability for
    non-root to read config values.

Change-Id: Ief197628d9d161276efb211930523fea674251d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2048345
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>