init: pre-startup: convert to task
This should be a task so Upstart doesn't return until it finishes
executing.  Otherwise it waits for the first fork which could be
too fast.  While unlikely, and doesn't seem to have happened in
practice that we've noticed, might as well have it be correct to
avoid confusing others reading the code.

BUG=None
TEST=CQ passes

Change-Id: Iff81131248a5ec07a9e36372c6df58d84d7bcb6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2048277
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>