crash: Add "early" metadata field.
This field will be populated by crashes that are collected in early
boot.

Right now, this is primarily useful to test early boot crash handling,
but it could also be useful to look at for metrics in the field and/or
to debug failures.

BUG=chromium:1039378
TEST=Manually induce a crash on an eve, verify that "early" was present in meta file iff crash was early.

Change-Id: I0c9858fe6f5119e02ea465e1879d1d37a961edb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2048869
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>