vm_tools: add support for image_type in ListVmDisks
Report the type of each disk in ListVmDisks.  This can be used to
determine if a disk can be resized (raw disk image) or not.

For completeness, plugin VM is added as a DiskImageType; this is
currently unused in CreateDiskImage, since plugin VMs are also in a
different location, and non-plugin VM creation already handles unknown
disk image types as an error condition.

BUG=chromium:858815
TEST=`vmc list` and verify type is as expected ('raw')

Change-Id: I9572465911a12b915647fdd3d4f15c693049dc77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2051159
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>