crash: Check return value of CreateCrashFormData.
This function can return NULL, but we weren't checking its return value.

NOTE: I wasn't able to make the reproducer _fail_ at all, so my
confidence that this fixes the linked issue is limited. My plan is to
wait and see if clusterfuzz marks the bug closed.

That said, we do need to be checking for null return values, and failure
to do that would cause a crash similar to what we see in this bug.

BUG=chromium:1048727
TEST=cros_fuzz  --board=eve reproduce --fuzzer crash_sender_fuzzer \
--testcase \
/build/eve/tmp/clusterfuzz-testcase-minimized-crash_sender_fuzzer-6293703608041472 \
--package crash-sender --build-type asan

Change-Id: I77d6d0aafc41919beb666f814b3079b6b9b93454
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2051312
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>