u2fd: Stop reporting WINK capability
Chrome did not use to issue the WINK command, and
instead relied on the device to do so itself when
presence was required; u2fd therefore displays a
notification when the user needs to press the power
button.

A change in Chrome means that it also issues the WINK
command to all connected U2F devices, for every
authenticate operation. This results in u2fd displaying
a notification asking the user to press the power button
even when the key is not owned by the integrated device.

By not reporting the WINK capability, Chrome does not
issue the command for the integrated device, and the UI
notifications are displayed as they were before the
Chrome change.

BUG=b:149185613
TEST=Register on webauthn.io with external key, log in
     and check that 'press power button' notification is
     shown once, and that logs reflect 1 WINK command.
     Build and deploy to soraka locally, repeat the above,
     check that no notification is displayed and no WINK
     is logged.

Change-Id: I7e569af845f2035b36e3ca911db643fcbfe0225f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2051786
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Louis Collard <louiscollard@chromium.org>