init: Create 'mnt_chrome' file.
After thinking about this for a while, this is the best place to
create this file.

The file itself does nothing, it just serves as an FS location to
bind-mount things onto. At the same time, we want the file to exist
independently of system service lifetime (the bind-mount, on the other
hand, will track Chrome/session_manager lifetime). So, just create it
in chromeos_startup.

An advantage of this is that this section can be the authoritative list
of all persistent namespace references.

BUG=chromium:985492
TEST=Deploy, reboot.
TEST=ls /run/namespaces/mnt_chrome

Change-Id: Ic2e70e2c3816a37354f37692fc6e17c0cb99d6af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2052343
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>