crash: Better logging for Chrome crashes if no consent
Currently, it is difficult to tell the difference in the logs between
"Chrome crashed but we ignored it because we didn't have consent" vs.
"Chrome crashed but never called crash_reporter". This is making it
difficult to debug certain tast failures.

In the breakpad days, we wouldn't expect to be called at all if consent
was off, so it was less of an issue. But crashpad will always call us;
it expects crash_reporter to handle consent issues.

BUG=chromium:1041062
TEST=Ran with no consent, saw message

Change-Id: I0fa364f22b3569c0f99a2c616fda70f45b545ad0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2055444
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Auto-Submit: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>