cros-disks: RarManager calls rar2fs with umask option
This sets the permissions on the files exposed in the mounted archive,
and avoids problems linked to recorded permissions, in particular when
dealing with nested archives (ie RAR in RAR).

Note that, since RarManager is not used by cros-disks yet, there is no
change in cros-disks's behavior. This is essentially preparing
RarManager to be a replacement of ArchiveManager.

BUG=chromium:996549
TEST=cros_workon_make cros-disks --test

Cq-Depend: chromium:2038215
Change-Id: I94329667b4094b38a86c75595acde19d53848fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2060535
Tested-by: François Degros <fdegros@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>